fix: restore production build boundaries
