+++ .trellis/tasks/08-07-publish-and-merge-architecture-cleanup/pr-draft.md
@@ -11,11 +11,12 @@
 - Consolidate dashboard/home benefit projection, subscription handling, and neutral AMEX catalog ownership behind canonical modules.
 - Remove superseded migration/search/analytics/UI utilities and stale product/docs assets.
 - Refresh the supported Next/SWC/ESLint/Prisma/Auth/Axios dependency set and lockfile within current major versions.
+- Enforce a client-safe dashboard contract boundary and keep helpers out of App Router route-module exports.
 
 ## Verification
 
-- `npm test -- --runInBand` — 81 suites passed; 758 tests passed, 1 skipped.
-- Strict TypeScript, public-DB invariant, card-template validation, AMEX userscript artifact checks, package-context discovery, lockfile graph, Markdown-link review, and `git diff --check` passed.
+- `npm test -- --runInBand` — 82 suites passed; 760 tests passed, 1 skipped.
+- Production Next build, strict TypeScript, changed-source ESLint, public-DB invariant, card-template validation, AMEX userscript artifact checks, package-context discovery, lockfile graph, Markdown-link review, and `git diff --check` passed.
 - Dependency audit reduced from 30 findings to 6; remaining findings are documented deferred Next/Auth major-upgrade paths in the archived dependency-remediation evidence.
 
 ## Risks and operational skips
